правки

+++ src/main/java/com/home/dao/CommentDao.java
@@ -3,6 +3,6 @@ package com.home.dao;
 import com.home.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentDao extends JpaRepository<Comment, Integer> {
+public interface CommentDao extends JpaRepository<Comment, Long> {
 
 }
